Add `private` flag for Projects
Sometimes you may end up in a situation where you want both private and
public projects on your (self-hosted) Hydra. This is not doable with the
`hidden`-flag since this may hide your project from the list on the
landing page, but everything else is still accessible via both a
webbrowser and the REST API, so guessing relevant URLs is sufficient.

This is implemented here by adding a `private`-column to each project
that can be checked on creation or update. As soon as this has happened,
every request to something related to the project (i.e. jobsets, builds,
scm diff & channels) will be rejected with a 403 and the project and
corresponding builds will be hidden from overview pages, the search and
the landing page.

The only downside this currently has is that it's not possible to make
cached build results public/private if these appear in both a public and
private project.