Revert "Return 403 for NARs provided by Hydra's own binary cache if it is part of a private project"
This reverts commit 79768e4b0435d39e576fce9f921f04677dadbacb.

This is probably a bad idea. I discussed this with Graham directly and
it has to be sufficient to assume hashes to be unguessable. If somebody
wants 100% safety, it's probably better to set up a binary cache with
actual authentication.

Keeping this change in the branch for now though in case we need it (or
a part of it) later.