Build: enable nix-command on `nix path-info` call, silence stdout as well
`nix-store -q` (and `--outputs`) doesn't seem to complain if a path
doesn't exist in a store anymore, so let's do it that way.

This approach works fine on Nix 2.3 as well, it only warns about
`experimental-features` being an unknown option, but that's silenced
along with the rest of stderr.

(cherry picked from commit 15b7d91b512ce565319ecbf6c6fa63f408ea41a9)